move diff localnet detection logic from the flowalerts_analyzer interface to a common utils.py file
